requirements.txt
# Web Framework
fastapi==0.121.0
uvicorn[standard]==0.38.0

# Database
SQLAlchemy==2.0.44
psycopg2-binary==2.9.11

# Data Science & Machine Learning
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.7.2
scipy==1.16.3
xgboost==2.0.3
joblib==1.5.2

# Data Visualization (pour l'entraînement du modèle)
matplotlib==3.10.7
seaborn==0.13.2

# HTTP Requests
requests==2.32.5

# Task Scheduling
APScheduler==3.11.1

# SMS Notifications
twilio==9.8.5

# Environment & Configuration
python-dotenv==1.2.1

# Data Validation (dépendance de FastAPI, mais explicitement utilisée)
pydantic==2.12.4
